Esri Data replication from different data sources
Have you ever struggled with ArcGIS server data sources or tried to join SQL table that is not in the same Geo-database as the layer it should be joined to? This little script solves both of those problems. Assuming  you run ArcGIS REST Web services to serve your data out and have a Geo-database set up as a data source for this server. First  you want to copy data to this server and setup join and other dependencies in .mxd, Then you want to publish the service on your server. This creates a schema lock on the data. This script stops this service to unlock the schema, deletes your old data sets, and copies the new data sets from their sources. Then it starts the service again, and your data is updated. You can set it up to run as a Task on windows or as a chron-job on Linux.

I know it is very primitive way, but it works as a charm. 
Feel free to fork this code and improve it in any way.